+++ kraftwerk-api/src/main/java/fr/insee/kraftwerk/api/services/StepByStepService.java
@@ -46,6 +46,7 @@ public class StepByStepService extends KraftwerkService {
 			)  {
 		//Read data files
 		Path inDirectory;
+		boolean withDdi = true;
 		try {
 			inDirectory = controlInputSequence.getInDirectory(inDirectoryParam);
 		} catch (KraftwerkException e) {
@@ -67,7 +68,7 @@ public class StepByStepService extends KraftwerkService {
 		for (String dataMode : userInputs.getModeInputsMap().keySet()) {
 			VtlBindings vtlBindings = new VtlBindings();
 			try {
-				buildBindingsSequence.buildVtlBindings(userInputs, dataMode, vtlBindings,metadataVariables);
+				buildBindingsSequence.buildVtlBindings(userInputs, dataMode, vtlBindings,metadataVariables, withDdi );
 			} catch (NullException e) {
 				return ResponseEntity.status(e.getStatus()).body(e.getMessage());
 			}
@@ -90,6 +91,8 @@ public class StepByStepService extends KraftwerkService {
 			)  {
 		//Read data files
 		Path inDirectory;
+		boolean withDdi = true;
+
 		try {
 			inDirectory = controlInputSequence.getInDirectory(inDirectoryParam);
 		} catch (KraftwerkException e) {
@@ -108,7 +111,7 @@ public class StepByStepService extends KraftwerkService {
 		//Process
 		BuildBindingsSequence buildBindingsSequence = new BuildBindingsSequence(withAllReportingData);
 		try {
-			buildBindingsSequence.buildVtlBindings(userInputs, dataMode, vtlBindings, metadataVariables);
+			buildBindingsSequence.buildVtlBindings(userInputs, dataMode, vtlBindings, metadataVariables, withDdi);
 		} catch (NullException e) {
 			return ResponseEntity.status(e.getStatus()).body(e.getMessage());
 		}