+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/metadata/MetadataUtils.java
@@ -60,4 +60,15 @@ public class MetadataUtils {
 		}
 		variables.putVariable(new Variable(missingVar, group, varType));
 	}
+
+	public static Map<String, VariablesMap> getMetadataFromLunatic(Map<String, ModeInputs> modeInputsMap) {
+		Map<String, VariablesMap> metadataVariables = new LinkedHashMap<>();
+		modeInputsMap.forEach((k, v) -> putToMetadataVariableFromLunatic(k,v,metadataVariables));
+		return metadataVariables;
+	}
+
+	private static void putToMetadataVariableFromLunatic(String dataMode, ModeInputs modeInputs, Map<String, VariablesMap> metadataVariables ) {
+		VariablesMap variables = LunaticReader.getVariablesFromLunatic(modeInputs.getLunaticFile());
+		metadataVariables.put(dataMode, variables);
+	}
 }